fix: change types of Car object

+++ adapters/src/test/java/br/com/marcia/adapters/inbound/controllers/request/builder/CarRequestBuilder.java
@@ -16,7 +16,7 @@ public class CarRequestBuilder {
     private String manufacturer = "Automobili Lamborghini S.P.A.";
 
     @Builder.Default
-    private Long year = 2021L;
+    private Integer year = 2021;
 
     @Builder.Default
     private String placa = "OPA0148";
