+++ adapters/src/test/java/br/com/marcia/adapters/inbound/controllers/response/builder/CarResponseBuilder.java
@@ -20,7 +20,7 @@ public class CarResponseBuilder {
     private String manufacturer = "Automobili Lamborghini S.P.A.";
 
     @Builder.Default
-    private Long year = 2021L;
+    private Integer year = 2021;
 
     @Builder.Default
     private String placa = "OPA0148";
